add validation of form

+++ database/migrations/2022_08_24_223919_create_big_integer.php
@@ -15,7 +15,7 @@ class CreateBigInteger extends Migration
     {
         Schema::table('product', function (Blueprint $table) {
             //
-            $table->bigInteger('user_id')->change();
+            $table->Integer('user_id')->change();
 
         });
     }
